requirements.txt
asgiref==3.11.0
Django==5.2.8
django-allauth==65.13.0
django-ckeditor==6.7.3
django-environ==0.12.0
django-js-asset==3.1.2
pillow==12.0.0
sqlparse==0.5.3
requests==2.32.3
PyJWT==2.10.1
cryptography==44.0.0
